+++ examples/basic-ui-server/homebridge-ui/server.js
@@ -1,39 +1,37 @@
-const { HomebridgePluginUiServer, RequestError } = require('@homebridge/plugin-ui-utils');
-const { createHash } = require('crypto');
+import { createHash } from 'node:crypto'
+import { HomebridgePluginUiServer, RequestError } from '@homebridge/plugin-ui-utils'
 
 class PluginUiServer extends HomebridgePluginUiServer {
   constructor() {
     // super() MUST be called first
-    super();
+    super()
 
     // handle request for the /token route
-    this.onRequest('/token', this.generateToken.bind(this));
+    this.onRequest('/token', this.generateToken.bind(this))
 
     // this MUST be called when you are ready to accept requests
-    this.ready();
+    this.ready()
   }
 
   async generateToken(payload) {
-    console.log('Username:', payload.username);
+    // eslint-disable-next-line no-console
+    console.log('Username:', payload.username)
 
     // sleep for 1 second, just to demo async works
-    await new Promise((resolve) => setTimeout(resolve, 1000));
+    await new Promise(resolve => setTimeout(resolve, 1000))
 
     try {
       // generate a sha256 from the username and use that as a fake token
-      const hashedUsername = createHash('sha256').update(payload.username).digest().toString('hex');
+      const hashedUsername = createHash('sha256').update(payload.username).digest().toString('hex')
 
       // return data to the ui
       return {
         token: hashedUsername,
       }
     } catch (e) {
-      throw new RequestError('Failed to Generate Token', { message: e.message });
+      throw new RequestError('Failed to Generate Token', { message: e.message })
     }
-
   }
 }
 
-(() => {
-  return new PluginUiServer();
-})();
+(() => new PluginUiServer())()