+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to `@homebridge/plugin-ui-utils` will be documented in this 
 
 ### Notable Changes
 
-- update to esm package
+- ⚠️ update to esm package
+- add methods for enabling and disabling save button
+  - requires homebridge ui `^5.0.0-beta.4`
 
 ## v1.0.3 (2024-04-06)